Move peeled tags protocol functions to dulwich.protocol
Remove InfoRefsContainer class, inlining its functionality into
SwiftInfoRefsContainer in dulwich.contrib.swift.

This properly isolates the ^{} protocol syntax to protocol-level code,
which is a bit of an architectural leak.

Fixes #2009
