+++ NEWS
@@ -1,5 +1,12 @@
 0.25.0	UNRELEASED
 
+ * Move protocol-level peeled tags functions (``serialize_refs()``,
+   ``write_info_refs()``, ``split_peeled_refs()``, ``strip_peeled_refs()``)
+   from ``dulwich.refs`` to ``dulwich.protocol``. The ``^{}`` peeled tags syntax
+   is now properly isolated to protocol-level code. Remove ``InfoRefsContainer``
+   class (functionality inlined into ``SwiftInfoRefsContainer``).
+   (Jelmer Vernooĳ, #2009)
+
  * Fix ``get_unstaged_changes()`` to correctly pass Blob objects to filter
    callbacks instead of raw bytes. This fixes crashes when using ``.gitattributes``
    files with filter callbacks like ``checkin_normalize``.
