+++ tests/test_refs.py
@@ -30,9 +30,9 @@ from typing import ClassVar
 from dulwich import errors
 from dulwich.file import GitFile
 from dulwich.objects import ZERO_SHA
+from dulwich.protocol import split_peeled_refs, strip_peeled_refs
 from dulwich.refs import (
     DictRefsContainer,
-    InfoRefsContainer,
     NamespacedRefsContainer,
     SymrefLoop,
     _split_ref_line,
@@ -43,8 +43,6 @@ from dulwich.refs import (
     read_packed_refs,
     read_packed_refs_with_peeled,
     shorten_ref_name,
-    split_peeled_refs,
-    strip_peeled_refs,
     write_packed_refs,
 )
 from dulwich.repo import Repo
@@ -1079,55 +1077,6 @@ class DiskRefsContainerPathlibTests(TestCase):
         # Test refpath returns worktree path for HEAD
         ref_path = refs.refpath(b"HEAD")
         self.assertEqual(ref_path, os.path.join(worktree_dir.encode(), b"HEAD"))
-
-
-class InfoRefsContainerTests(TestCase):
-    def test_invalid_refname(self) -> None:
-        text = _TEST_REFS_SERIALIZED + b"00" * 20 + b"\trefs/stash\n"
-        refs = InfoRefsContainer(BytesIO(text))
-        expected_refs = dict(_TEST_REFS)
-        del expected_refs[b"HEAD"]
-        expected_refs[b"refs/stash"] = b"00" * 20
-        del expected_refs[b"refs/heads/loop"]
-        self.assertEqual(expected_refs, refs.as_dict())
-
-    def test_keys(self) -> None:
-        refs = InfoRefsContainer(BytesIO(_TEST_REFS_SERIALIZED))
-        actual_keys = set(refs.keys())
-        self.assertEqual(set(refs.allkeys()), actual_keys)
-        expected_refs = dict(_TEST_REFS)
-        del expected_refs[b"HEAD"]
-        del expected_refs[b"refs/heads/loop"]
-        self.assertEqual(set(expected_refs.keys()), actual_keys)
-
-        actual_keys = refs.keys(b"refs/heads")
-        actual_keys.discard(b"loop")
-        self.assertEqual(
-            [b"40-char-ref-aaaaaaaaaaaaaaaaaa", b"master", b"packed"],
-            sorted(actual_keys),
-        )
-        self.assertEqual([b"refs-0.1", b"refs-0.2"], sorted(refs.keys(b"refs/tags")))
-
-    def test_as_dict(self) -> None:
-        refs = InfoRefsContainer(BytesIO(_TEST_REFS_SERIALIZED))
-        # refs/heads/loop does not show up even if it exists
-        expected_refs = dict(_TEST_REFS)
-        del expected_refs[b"HEAD"]
-        del expected_refs[b"refs/heads/loop"]
-        self.assertEqual(expected_refs, refs.as_dict())
-
-    def test_contains(self) -> None:
-        refs = InfoRefsContainer(BytesIO(_TEST_REFS_SERIALIZED))
-        self.assertIn(b"refs/heads/master", refs)
-        self.assertNotIn(b"refs/heads/bar", refs)
-
-    def test_get_peeled(self) -> None:
-        refs = InfoRefsContainer(BytesIO(_TEST_REFS_SERIALIZED))
-        # refs/heads/loop does not show up even if it exists
-        self.assertEqual(
-            _TEST_REFS[b"refs/heads/master"],
-            refs.get_peeled(b"refs/heads/master"),
-        )
 
 
 class ParseSymrefValueTests(TestCase):
